+++ worker/src/handles/Timer.cpp
@@ -56,7 +56,7 @@ void Timer::Start(uint64_t timeout)
 
 	int err;
 
-	err = uv_timer_start(this->uvHandle, (uv_timer_cb)on_timer, (uint64_t)timeout, 0);
+	err = uv_timer_start(this->uvHandle, (uv_timer_cb)on_timer, timeout, 0);
 	if (err)
 		MS_THROW_ERROR("uv_timer_start() failed: %s", uv_strerror(err));
 }
